doc: Delimit Map types with dash
@param or @prop with a type of Map<keyType,valuType> does not
render correctly. It fractures just after the delimiter. The
second type and closing angle bracket end up at the beginning of
the line, before the object name. Unsure whether this is a JSDoc
or sphinx-js effect.

@return with a type of the same pattern renders correctly.

Delimits with dash - instead of comma , as a hack to get a
sensible rendering.